Fix deferred section hydration in client components
